# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pompote-view API",
    "description": "Read-only HTTP surface for monitoring trading strategies. Every `/api/v1/*` route requires the `X-API-Key` header.",
    "contact": {
      "name": "Pierre Attard"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/monitoring/backtests": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/backtests`.",
        "operationId": "list_backtests",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Optional `status` filter (`running`/`completed`/`failed`/`aborted`).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "strategy_kind",
            "in": "query",
            "description": "Optional `strategy_kind` filter.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "exchange",
            "in": "query",
            "description": "Optional `exchange` filter.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "description": "Optional `symbol` filter.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Row cap (defaults to and rejected above `MAX_BACKTEST_RUNS`).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Backtest runs ordered by descending `started_at`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BacktestRunSummaryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid `status` filter or `limit` out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/backtests/{run_id}": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/backtests/{run_id}`.",
        "operationId": "get_backtest",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "description": "Backtest run identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The run and its configuration snapshot.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestRunDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "404": {
            "description": "No run with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/backtests/{run_id}/candles": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/backtests/{run_id}/candles`.",
        "operationId": "get_backtest_candles",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "description": "Backtest run identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Aggregation width (must be one of [`Timeframe::ALLOWED`]).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated candles_5s over the run's market window, with a `source` hint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestCandlesDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid timeframe or window exceeding the bucket cap.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "404": {
            "description": "No run with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/backtests/{run_id}/decisions": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/backtests/{run_id}/decisions`.",
        "operationId": "get_backtest_decisions",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "description": "Backtest run identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `market_ts` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `market_ts` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Row cap (defaults to and rejected above `MAX_ORDER_ROWS`).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run decisions on the window, ordered by ascending `market_ts`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range or limit out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/backtests/{run_id}/fills": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/backtests/{run_id}/fills`.",
        "operationId": "get_backtest_fills",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "description": "Backtest run identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `market_ts` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `market_ts` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Row cap (defaults to and rejected above `MAX_ORDER_ROWS`).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run fills on the window, ordered by ascending `market_ts`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FillDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range or limit out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
...
    "/api/v1/monitoring/backtests/{run_id}/orders": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/backtests/{run_id}/orders`.",
        "operationId": "get_backtest_orders",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "description": "Backtest run identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `market_ts` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `market_ts` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Row cap (defaults to and rejected above `MAX_ORDER_ROWS`).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run orders on the window, ordered by ascending `market_ts`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BacktestOrderDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range or limit out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/candles": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/candles`.",
        "operationId": "get_candles",
        "parameters": [
          {
            "name": "exchange",
            "in": "query",
            "description": "Exchange identifier as stored in `candles_5s.exchange`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "description": "Symbol identifier as stored in `candles_5s.symbol`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Aggregation width (must be one of [`Timeframe::ALLOWED`]).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `open_time` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `open_time` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated OHLCV buckets ordered by ascending `ts`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandleDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid timeframe, range or too many points.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error (e.g. schema drift).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/strategies": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/strategies`.",
        "operationId": "list_strategies",
        "responses": {
          "200": {
            "description": "Every strategy, ordered by name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/strategies/{id}/decisions": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/strategies/{id}/decisions`.",
        "operationId": "get_strategy_decisions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Strategy identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `executed_at` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `executed_at` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Row cap (defaults to [`MAX_ORDER_ROWS`], rejected above it).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Live decisions (with market context) for the strategy on the window, ordered by ascending `created_at`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LiveDecisionDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range or limit out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
...
    "/api/v1/monitoring/timeframes": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/timeframes`.",
        "operationId": "get_timeframes",
        "responses": {
          "200": {
            "description": "Allowed aggregation timeframes (finest to coarsest).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "BacktestErrorBody": {
        "type": "object",
        "description": "JSON error body for the backtest endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "allowed": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Whitelisted values for `invalid_status`."
          },
          "error": {
            "type": "string",
            "description": "Machine-readable discriminator (e.g. `invalid_range`)."
          },
          "max": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Applicable row cap for `too_many_rows`.",
            "minimum": 0
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable explanation."
          },
          "requested": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Echoes the requested row count for `too_many_rows`.",
            "minimum": 0
          }
        }
      },
...
      "BacktestRunDetailDto": {
        "type": "object",
        "description": "A run with its opaque configuration snapshot (detail view).",
        "required": [
          "run",
          "config_snapshot"
        ],
        "properties": {
          "config_snapshot": {
            "type": "object",
            "description": "Opaque, display-only configuration snapshot (`jsonb` object)."
          },
          "run": {
            "$ref": "#/components/schemas/BacktestRunSummaryDto",
            "description": "The run summary."
          }
        }
      },
...
    },
    "securitySchemes": {
      "x_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}